LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/json-module/integrity-expected.txt

CONSOLE MESSAGE: Cannot load script http://localhost:8800/html/semantics/scripting-1/the-script-element/json-module/integrity-mismatches.js. Failed integrity metadata check. Content length: 120, Expected content length: 120, Expected metadata: sha384-doesnotmatch
CONSOLE MESSAGE: TypeError: Cannot load script due to integrity mismatch

PASS The integrity attribute must be verified on the top-level of a module loading a JSON module and allow it to execute when it matches
PASS The integrity attribute must be verified on the top-level of a module loading a JSON module and not allow it to execute when there's a mismatch
